requirements/base.txt
pylint==2.2.2
pytest==4.3.0
pyOpenSSL==19.0.0
passlib==1.7.1
psycopg2-binary==2.7.7
Django==1.11.16
django-extensions==2.1.5
django-werkzeug-debugger-runserver==0.3.1
django-allauth==0.38.0
django-sslserver==0.20
djangorestframework==3.9.1
firebase_admin==2.15.1
django-rest-framework-social-oauth2
django-cors-headers
psycopg2
gunicorn
dj-database-url
django-crispy-forms
pillow